moved main to single cpp file
