+++ src/hardwareio/IMUCalibDumper/IMUCalibDumper.cpp
@@ -10,9 +10,6 @@
  * @authors:    Leandro de Souza Rosa <leandro.desouzarosa@iit.it>
  */
 
-#ifndef __IMU_CALIB_DUMP_H
-#define __IMU_CALIB_DUMP_H
-
 #include <yarp/sig/Vector.h>
 #include <yarp/os/all.h>
 #include <event-driven/all.h>
@@ -211,5 +208,26 @@ public:
         }
     }
 };
-#endif
+
+int main(int argc, char * argv[])
+{
+    /* initialize yarp network */
+    yarp::os::Network yarp;
+    if(!yarp.checkNetwork(2)) {
+        std::cout << "Could not connect to YARP" << std::endl;
+        return -1;
+    }
+
+    /* prepare and configure the resource finder */
+    yarp::os::ResourceFinder rf;
+    rf.setVerbose( false );
+    rf.setDefaultContext( "event-driven" );
+    rf.setDefaultConfigFile( "IMUCalibDumper.ini" );
+    rf.configure( argc, argv );
+
+    /* create the module */
+    IMUCalibDumper instance;
+    return instance.runModule(rf);
+}
+
 //That empty line
